chore: move time manager to separate directory
Bench: 4959499
