+++ src/core/uci.cpp
@@ -20,7 +20,7 @@
 #include "search/search.h"
 #include "spsa/spsa.h"
 #include "tests/see_test.h"
-#include "time_manager.h"
+#include "time_manager/time_manager.h"
 #include "tt/tt.h"
 #include "utils/perft.h"
 #include "utils/str_utils.h"
